+++ src/LongBaseSequence.cpp
@@ -141,9 +141,9 @@ void ChanZuckerberg::shasta::testLongBaseSequence()
         LongBaseSequences sequences;
         sequences.createNew("abc", 4096);
 
-        for(size_t i=0; i<8; i++) {
+        for(uint8_t i=0; i<8; i++) {
             LongBaseSequence sequence(i*50);
-            const Base base = Base::fromInteger(i%4);
+            const Base base = Base::fromInteger(uint8_t(i%4));
             for(size_t j=0; j<sequence.baseCount; j++) {
                 sequence.set(j, base);
             }
@@ -151,10 +151,10 @@ void ChanZuckerberg::shasta::testLongBaseSequence()
             sequences.append(sequence);
         }
         cout << endl;
-        for(size_t i=0; i<sequences.size(); i++) {\
+        for(uint8_t i=0; i<sequences.size(); i++) {\
             cout << i << endl;
             LongBaseSequenceView sequence = sequences[i];
-            const Base base = Base::fromInteger(i%4);
+            const Base base = Base::fromInteger(uint8_t(i%4));
             for(size_t j=0; j<sequence.baseCount; j++) {
                 sequence.set(j, base);
             }
